remove core and merge missing files
