+++ src/chariot/network/NetworkManager.py
@@ -21,8 +21,7 @@ class NetworkManager:
             self.userNetworks: Dict[str, Network] = {}
             NetworkManager.__instance = self
 
-    # This method gives a new network name to an already defined network. Find
-    # the network by name
+    # This method gives a new network name to an already defined network. Find the network by name
     def modifyNetworkNameByName(self, newName: str, toFind: str):
         # check that newName is unique to avoid data overwrites
         if self.isNetworkNameUnique(newName) is False:
@@ -35,8 +34,7 @@ class NetworkManager:
             # update network instance
             network.modifyNetworkName(newName)
             # update collection, make new key and delete the old one
-            self.userNetworks[network.getNetworkName()
-                              ] = self.userNetworks[toFind]
+            self.userNetworks[network.getNetworkName()] = self.userNetworks[toFind]
             del self.userNetworks[toFind]
         else:
             # throw exception that network name is not found in collection
@@ -51,15 +49,13 @@ class NetworkManager:
             # throw exception that network name is not found in collection
             pass
 
-    # This method adds a Network object to the dictionary of Networks that
-    # this class manages
+    # This method adds a Network object to the dictionary of Networks that this class manages
     def addNetwork(self, networkName: str, networkDesc: str):
         if networkName is None:
             # raise expection here
             pass
 
-        # When adding a key, if the same key is in the dict, the values are
-        # overriden. Avoid this.
+        # When adding a key, if the same key is in the dict, the values are overriden. Avoid this.
         if self.isNetworkNameUnique(networkName):
             network = Network(networkName, networkDesc)
             self.userNetworks[networkName] = network
@@ -67,18 +63,15 @@ class NetworkManager:
             # network name is not unique, throw correct exception
             pass
 
-    # This method deletes a Network object via name from the nested dictionary
-    # of Networks that this class manages
+    # This method deletes a Network object via name from the nested dictionary of Networks that this class manages
     def deleteNetworkByName(self, netName: int):
         networkToDel: Network = self.findNetworkByNetworkName(netName)
         if networkToDel is not None:
             del self.userNetworks[networkToDel.getNetworkName()]
 
-    # Convenience method used to return a Network object via a given network
-    # name
+    # Convenience method used to return a Network object via a given network name
     def findNetworkByNetworkName(self, netNameToFind: str) -> Network:
-        # since the collection is nested dictionaries, find the name with
-        # .get()
+        # since the collection is nested dictionaries, find the name with .get()
         try:
             network: Network = self.userNetworks.get(netNameToFind)
         except KeyError:
@@ -86,8 +79,7 @@ class NetworkManager:
             pass
         return network
 
-    # Convenience method that checks whether a network by the same name exists
-    # in the collection
+    # Convenience method that checks whether a network by the same name exists in the collection
     def isNetworkNameUnique(self, uniqueName: str) -> bool:
         found: bool = False
         if self.findNetworkByNetworkName(uniqueName) is not None:
