--- src/chariot/test/mongo-db.py
@@ -1,5 +1,0 @@
-from chariot.database.MongoDatabase import MongoDatabase
-
-mongo = MongoDatabase()
-mongo.insertRow(100, "hi")
-print(__name__)
